+++ src/newconf.c
@@ -2320,6 +2320,7 @@ static struct ConfEntry conf_general_table[] =
 	{ "static_part_reason", CF_QSTRING, NULL, REALLEN, &ConfigFileEntry.static_part_reason },
         { "anti_spam_exit_message_time", CF_TIME,  NULL, 0, &ConfigFileEntry.anti_spam_exit_message_time },
 	{ "use_part_messages",		CF_YESNO, NULL, 0, &ConfigFileEntry.use_part_messages	},
+	{ "hide_uids_in_whois", CF_YESNO, NULL, 1, &ConfigFileEntry.hide_uids_in_whois },
 	{ "disable_fake_channels",	 CF_YESNO, NULL, 0, &ConfigFileEntry.disable_fake_channels },
 	{ "hide_channel_below_users", CF_INT, NULL, 0, &ConfigFileEntry.hide_channel_below_users },
 	{ "min_nonwildcard_simple",	 CF_INT,   NULL, 0, &ConfigFileEntry.min_nonwildcard_simple },